refactor SelfMerge & unit tests
